+++ 0x08-recursion/main.h
@@ -8,6 +8,6 @@ void _print_rev_recursion(char *s);
 int _strlen_recursion(char *s);
 int factorial(int n);
 int _pow_recursion(int x, int y);
-int _sqrt_recursion(int n);
-int _sqrt_recursion2(int n, int i, int j);
+int _sqrt_recursion(long int n);
+int _sqrt_recursion2(long int n, long int i, long int j);
 #endif /* MAIN_H */
